New pathing
This is an overview of the remaining work.
