+++ simple-rest-server/main.go
@@ -14,25 +14,6 @@ import (
 	"github.com/julienschmidt/httprouter"
 )
 
-func getIndexByTaskID(taskID int) (index int, err error) {
-	// If we didn't find the task ID - return an error and set the index to -1...
-	err = errors.New("GetIndexByTaskID: TaskID " + string(taskID) + " Not Found.")
-	index = -1
-
-	accessTasks.Lock()
-
-	// Search for the taskID and return the index.
-	for i, line := range allTasks {
-		if taskID == line.ID {
-			index = i
-			err = nil
-			break
-		}
-	}
-	accessTasks.Unlock()
-	return
-}
-
 type Task struct {
 	ID        int
 	Checked   bool      `json:"checked"`
@@ -80,18 +61,22 @@ func main() {
 	go func() {
 		for {
 			saveCSV()
-			time.Sleep(1 * time.Minute)
+			time.Sleep(5 * time.Second)
 		}
 	}()
 
 	// Start the API
 	router := httprouter.New()
-	router.GET("/", Index)
-	router.GET("/search", SearchTask)
-	router.GET("/list", ListTask)
-	router.POST("/add", AddTask)
-	router.POST("/checkoff", CheckOff)
-	router.DELETE("/delete", DeleteTask)
+
+	router.GET("/v1/task", ListTask)
+	router.POST("/v1/task", AddTask)
+	router.PUT("/v1/task", nil)
+	router.DELETE("/v1/task", nil)
+
+	router.GET("/v1/task/:id", SearchTask)
+	router.POST("/v1/task/:id", nil)
+	router.PUT("/v1/task/:id", Modify)
+	router.DELETE("/v1/task/:id", DeleteTask)
 
 	log.Fatal(http.ListenAndServe(":8080", router))
 }
@@ -113,4 +98,23 @@ func ifPanic(err error) {
 		fmt.Println(err)
 		os.Exit(1)
 	}
+}
+
+func getIndexByTaskID(taskID int) (index int, err error) {
+	// If we didn't find the task ID - return an error and set the index to -1...
+	err = errors.New("GetIndexByTaskID: TaskID " + string(taskID) + " Not Found.")
+	index = -1
+
+	accessTasks.Lock()
+
+	// Search for the taskID and return the index.
+	for i, line := range allTasks {
+		if taskID == line.ID {
+			index = i
+			err = nil
+			break
+		}
+	}
+	accessTasks.Unlock()
+	return
 }
