+++ simple-rest-server/post.go
@@ -3,10 +3,10 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/julienschmidt/httprouter"
 	"io/ioutil"
-	"log"
 	"net/http"
+
+	"github.com/julienschmidt/httprouter"
 )
 
 func AddTask(respWriter http.ResponseWriter, request *http.Request, _ httprouter.Params) {
@@ -19,7 +19,6 @@ func AddTask(respWriter http.ResponseWriter, request *http.Request, _ httprouter
 	if err != nil {
 		fmt.Println(err)
 	}
-	fmt.Fprintf(respWriter, "Show Details for ToDo '%v'\n", task)
 
 	accessTasks.Lock()
 	task.ID = allTasks[len(allTasks)-1].ID + 1
@@ -28,42 +27,4 @@ func AddTask(respWriter http.ResponseWriter, request *http.Request, _ httprouter
 	accessTasks.Unlock()
 
 	saveCSV()
-}
-
-type Check struct {
-	ID int `json:"id"`
-}
-
-func CheckOff(respWriter http.ResponseWriter, request *http.Request, _ httprouter.Params) {
-	var check Check
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		log.Println(err)
-		return
-	}
-	err = json.Unmarshal(body, &check)
-	if err != nil {
-		log.Println(err)
-		return
-	}
-
-	checkTaskAsComplete(check.ID)
-}
-
-func checkTaskAsComplete(taskID int) (err error) {
-	// Get the index.
-	index, err := getIndexByTaskID(taskID)
-	if err != nil {
-		log.Println(err)
-		return
-	}
-
-	// Lock the tasks.
-	accessTasks.Lock()
-
-	// Set the task to checked.
-	allTasks[index].Checked = true
-
-	accessTasks.Unlock()
-	return
 }
